+++ Domain/Domain.http
@@ -1,0 +1,6 @@
+@Domain_HostAddress = http://localhost:5230
+
+GET {{Domain_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
