+++ Infrastructure/Infrastructure.http
@@ -1,0 +1,6 @@
+@Infrastructure_HostAddress = http://localhost:5211
+
+GET {{Infrastructure_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
