--- LibraryApp.http
@@ -1,6 +1,0 @@
-@LibraryApp_HostAddress = http://localhost:5253
-
-GET {{LibraryApp_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
