+++ Presentation/Controllers/BooksController.cs
@@ -1,20 +1,21 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
 
-namespace LibraryApp.Api.Controllers
+namespace Api.Controllers
 {
     [ApiController]
     [Route("[controller]")]
-    public class UserController : ControllerBase
+    public class BooksController : ControllerBase
     {
+
         private readonly ILogger<TestController> _logger;
 
-        public UserController(ILogger<TestController> logger)
+        public BooksController(ILogger<TestController> logger)
         {
             _logger = logger;
         }
 
-        [HttpGet(Name = "GetUsers")]
+        [HttpGet(Name = "GetBooks")]
         public bool Get()
         {
             return true;
