+++ Presentation/Presentation.http
@@ -1,0 +1,6 @@
+@Presentation_HostAddress = http://localhost:5253
+
+GET {{Presentation_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
