+++ Presentation/WeatherForecast.cs
@@ -1,4 +1,4 @@
-namespace LibraryApp
+namespace Presentation
 {
     public class WeatherForecast
     {
