+++ Tests/Tests.csproj
@@ -1,0 +1,47 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <Compile Remove="Application\**" />
+	  <Compile Remove="Domain\**" />
+	  <Compile Remove="Frontend\**" />
+	  <Compile Remove="IAC\**" />
+	  <Compile Remove="Infrastructure\**" />
+	  <Content Remove="Application\**" />
+	  <Content Remove="Domain\**" />
+	  <Content Remove="Frontend\**" />
+	  <Content Remove="IAC\**" />
+	  <Content Remove="Infrastructure\**" />
+	  <EmbeddedResource Remove="Application\**" />
+	  <EmbeddedResource Remove="Domain\**" />
+	  <EmbeddedResource Remove="Frontend\**" />
+	  <EmbeddedResource Remove="IAC\**" />
+	  <EmbeddedResource Remove="Infrastructure\**" />
+	  <None Remove="Application\**" />
+	  <None Remove="Domain\**" />
+	  <None Remove="Frontend\**" />
+	  <None Remove="IAC\**" />
+	  <None Remove="Infrastructure\**" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+		<PackageReference Include="Moq" Version="4.20.72" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
+		<PackageReference Include="xunit" Version="2.9.2" />
+		<PackageReference Include="xunit.runner.visualstudio" Version="3.0.0">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Presentation\Presentation.csproj" />
+	</ItemGroup>
+
+</Project>
