+++ Tests/Tests.http
@@ -1,0 +1,6 @@
+@Tests_HostAddress = http://localhost:5117
+
+GET {{Tests_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
